RNTester: Add missing key attribute in NewAppScreen example (#27180)
Summary:
This pull request adds a missing key attribute to an array of elements in the example for `NewAppScreen`. This results in the "missing key" warning no longer appearing when viewing this example.

## Changelog

[General] [Fixed] - RNTester: Add missing key attribute in NewAppScreen example
Pull Request resolved: https://github.com/facebook/react-native/pull/27180

Test Plan: When viewing this example in RNTester, this warning is no longer printed.

Differential Revision: D18418198

Pulled By: mdvacca

fbshipit-source-id: 8fbe51da7b714e3e94fd2186ea8cc4e3606d9f95