Remove outdated references in `react-native/package.json`
Summary:
These files or directories no longer exist, so they don't need to be listed as files to be published to npm. The only file that was removed that does exist is the `.flowconfig` file, which should not actually be shipped to npm as projects using Flow would have their own `.flowconfig` file.

Changelog: [Internal]

Reviewed By: GijsWeterings

Differential Revision: D23132479

fbshipit-source-id: dc656d8215c9dac11025902b403c3f6560e62e37